Block delayed (#5)
New experimental group_by primitives and various bug fixes

- New experimental interface for group_by_key, histogram_by_key, and reduce_by_key
- New experimental interface for group_by_index, histogram_by_index, and reduce_by_index
- New experimental feature: block delayed sequences
- Add operator< and std::hash to sequences
- parlay::sequence is now short by default for T = char to fix some interface issues with character strings

+++ include/parlay/internal/scheduler_plugins/sequential.h
@@ -16,7 +16,7 @@ inline void set_num_workers(int) { ; }
 
 template <class F>
 inline void parallel_for(size_t start, size_t end, F f,
-                         size_t, bool) {
+                         long, bool) {
   for (size_t i=start; i<end; i++) {
     f(i);
   }
